Got a working shader running.
Changelog:
Shader working - added lots of shader code
Jenkins - Changed how CI handles examples. Now all are in one tar.gz file

+++ Scripts/Linux-Build.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
 
+while getopts "da" flag; do
+	case "${flag}" in
+	  d) dodoxygen="TRUE";;
+	  a) doarchive="TRUE";;
+  esac
+done
+
+
 # Setup
 echo "Ensuring Submodules are cloned."
 git submodule update --init --recursive || exit
 
 # CMake
 echo "Configuring CMake."
+rm -rf Build || exit
 mkdir -p Build
 cd Build || exit
 cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=../_install .. || exit
@@ -15,17 +24,40 @@ echo "Configuring CMake done."
 echo "Compiling the engine and example meta package"
 make Matts_Game_Engine-Examples
 echo "Engine and example meta package compiled"
-mkdir -p bin || exit
-cp Sources/*/lib* bin/ || exit
-cp Sources/Examples/*/Example* bin/ || exit
 
+#mkdir -p bin || exit
+#cp Sources/*/lib* bin/ || exit
+#cp Sources/Examples/*/Example* bin/ || exit
 
-# Doxygen
-echo "Cleaning Old Doxygen"
-rm -rf ../Docs || exit
-echo "Making Doxygen"
-make Matts_Game_Engine-Doxygen
+if [ "$dodoxygen" == "TRUE" ]; then
+  # Doxygen
+  echo "Cleaning Old Doxygen"
+  rm -rf ../Docs || exit
+  echo "Making Doxygen"
+  make Matts_Game_Engine-Doxygen
 
-# Archive Doxygen
-echo "Archive Doxygen"
-tar cvzf bin/Docs.tgz ../Docs
+  # Archive Doxygen
+  echo "Archive Doxygen"
+  tar cvzf bin/Docs.tar.gz ../Docs
+fi
+
+if [ "$doarchive" == "TRUE" ]; then
+  echo "Archiving Artifacts"
+  mkdir -p bin/artifacts
+
+  echo "Archiving Libraries"
+  cd bin || exit
+  tar cvzf artifacts/Libraries.tar.gz lib*
+  cd ..
+
+  echo "Archiving Doxygen"
+  if [ -e bin/Docs.tar.gz ]
+    then
+    cp bin/Docs.tar.gz bin/artifacts
+  fi
+
+  echo "Archiving Examples"
+  cd bin/Examples || exit
+  tar -cvzf "../artifacts/Examples.tar.gz" *;
+  cd ../..
+fi
